requirements.txt
pygrphviz
django-extensions
psycopg2